Add registry feature
