+++ Cargo.toml
@@ -11,4 +11,8 @@ keywords = ["macro", "field", "path", "accessor", "registry"]
 categories = ["no-std", "rust-patterns"]
 
 [dependencies]
-hashbrown = { version = "0.16", default-features = false, features = ["default-hasher", "inline-more"] }
+hashbrown = { version = "0.16", default-features = false, features = ["default-hasher", "inline-more"], optional = true }
+
+[features]
+default = ["registry"]
+registry = ["dep:hashbrown"]
